% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOvcov.R
\name{LOvcov}
\alias{LOvcov}
\title{Variance-covariance estimator that is robust to heteroskedasticity and many regressors}
\usage{
LOvcov(linmod, V)
}
\arguments{
\item{linmod}{an object of class "lm" that stems from fitting the regression model \eqn{y = X\beta + \epsilon}.}

\item{V}{a matrix specifying the linearly independent restrictions on the parameters.
The number of rows in \code{V} is equal to the number of contrasts and the number of columns in \code{V} must be equal to the dimension of \eqn{\beta}.}
}
\value{
\code{LOvcov} returns a variance-covariance matrix for the linear contrasts \eqn{V\beta}, using
the variance estimator proposed in Kline, Saggio, and Sølvsten (2019).
}
\description{
Computes a variance-covariance estimate for a set of linear contrasts \eqn{V\beta} in the regression model \eqn{y = X\beta + \epsilon} using
the leave-out estimator from Kline, Saggio, and Sølvsten (2019).  The leave-out estimator
provides a variance estimator which is asymptotically valid under many regressors and
unrestricted heteroskedasticity.
}
\examples{
## An example of a regression with 640 observations and 512 regressors
## where the last three coefficients are of interest.
set.seed(1)
X <- cbind(1, (0.5+runif(640))*matrix(exp(rnorm(640*511)), 640,511))
y <- X \%*\% c( -.5, rep(0.002,511) ) + rnorm(640)*.3*rowMeans(X)^2
V <- cbind( matrix(0, 3, 509), diag(3))
linmod <- lm(y ~ X-1)

vcov <- LOvcov(linmod, V)

## Coefficients and standard errors of interest
coeff <- V \%*\% linmod$coefficients
cbind(coeff,sqrt(diag(vcov)))

## P-value for joint significance test
1-pchisq( t(coeff) \%*\% solve(vcov) \%*\% coeff, df=3)
}
\references{
Kline, Saggio, and Sølvsten (2019). \emph{Leave-out estimation of variance components}. \url{https://arxiv.org/abs/1806.01494}
}
